Enhancement of Configuration Structure and bug fixes.
- A configuration file now has the capability to be divided
into fragment files. Each fragment mirrors the properties
found in the main config.json file. Any file identified
in the root folder of the configuration directory with
the ".frag" suffix is recognized as a fragment file and
is combined into a unified configuration file.

This enhancement does not replace the requirement for
the main config.json file, which must still be present
in the configuration folder when initializing the framework.

- Bug fixes.

Signed-off-by: Luis Silva <luis.m.silva99@hotmail.com>